+++ __test__/filechanger.test/src/index2.js
@@ -1,0 +1,3 @@
+module.exports = {
+    index2: "index2 page"
+};