+++ src/index.ts
@@ -1,0 +1,2 @@
+import FileChanger from "./api/FileChanger";
+module.exports = FileChanger;