+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-1.0.0 / 2016-03-15
-[#1]: https://github.com/kbukum/webpack-file-changer        "initial commit"
+# Change Log
+
+## 2.0.0
+   * changed project as typescript project.
+   * added new features like hash or hash:id usage to replace files content.
+   * added new trigger function
+     - before: called before operation. takes
+     - after: called after operation
+     - complete: called after complete all operations.
+